refactor(settings): split monolith into per-panel components, fix a11y and UX issues
Extract EmailSettings, BackupSettings, GitHubSettings, ServicesSettings into
separate files under app/(app)/settings/system/. Add shared useSystemSetting
hook to reduce duplicated fetch/save boilerplate.

Fixes addressed:
- #3: Unique IDs for Mailpace (sys-mailpace-apiKey) and Resend (sys-resend-apiKey)
- #4: Cancel button on each secret field restores masked display value
- #5: aria-label on Select triggers, Label htmlFor on Switch toggles
- #6: Edit buttons include field context (aria-label="Edit SMTP password", etc.)
- #7: Save button keeps text visible alongside spinner, adds aria-label
- #8: Monolith split into four focused components + shared hook
- #9: Remove dead ternary (editingSmtpPass ? smtpPass : smtpPass)
- #11: Provider/type changes reset provider-specific fields
- #12: autocomplete attributes on username, password, email, url fields
- #13: Loading spinners wrapped in role="status" with aria-live="polite"
